pyproject.toml
[tool.poetry]
name = "django_shop"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
Pillow = "^8.4.0"
djangorestframework = "^3.13.0"
drf-spectacular = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
